@@ -3,7 +3,7 @@ package th.weixia.yl.Akyuu.dto;
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 @Data
 public class MediaRequest {
