@@ -1,7 +1,7 @@
 package th.weixia.yl.Akyuu.dto;
 
 import lombok.Data;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 @Data
 public class MemberRequest {
